Finished pricelist sharing. Prices are editable. You also select a pricelist when you're in takeoff-mode.

+++ app/routes/_dashboard+/pricelists+/$pricelistId_.delete.tsx
@@ -1,16 +1,33 @@
-import { type ActionFunctionArgs, redirect } from "@remix-run/node";
-import { requireUserId } from "#app/utils/auth.server.js";
-import { prisma } from "#app/utils/db.server.js";
+import { invariantResponse } from '@epic-web/invariant'
+import { type ActionFunctionArgs, redirect } from '@remix-run/node'
+import { requireUserId } from '#app/utils/auth.server.js'
+import { prisma } from '#app/utils/db.server.js'
+import { requireUserWithPermission } from '#app/utils/permissions.server.js'
 
 export async function action({ params, request }: ActionFunctionArgs) {
-    // TODO: make sure the user has permission to delete this pricelist
-    await requireUserId(request);
+	const userId = await requireUserId(request)
+	const id = params.pricelistId
 
-    await prisma.pricelist.delete({
-        where: {
-            id: params.pricelistId,
-        },
-    });
+	const pricelist = await prisma.pricelist.findFirst({
+		where: {
+			id: id,
+		},
+	})
 
-    return redirect("/pricelists");
+	invariantResponse(pricelist, 'Not found', { status: 404 })
+
+	const isOwner = pricelist.ownerId === userId
+
+	await requireUserWithPermission(
+		request,
+		isOwner ? `delete:pricelist:own` : `delete:pricelist:any`,
+	)
+
+	await prisma.pricelist.delete({
+		where: {
+			id: id,
+		},
+	})
+
+	return redirect('/pricelists')
 }
